@@ -1,103 +1,76 @@
 {
-  "name": "basevite-blank",
-  "version": "0.0.1",
-  "description": "",
+  "name": "fake-tinder",
+  "author": "Filipe Petitemberte",
+  "version": "2.0.0",
+  "description": "Muito útil...",
   "main": "index.js",
   "scripts": {
-    "start": "concurrently \"npm:dev\" \"npm:gql-watch-schema\"",
-    "dev": "vite",
+    "start": "vite",
     "build": "tsc && vite build --mode production",
     "lint": "npm run check-types && eslint ./src",
     "lint-fix": "eslint ./src --fix",
     "check-types": "tsc --noemit",
-    "serve": "vite preview",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook -o ./dist-storybook",
-    "serve-storybook": "npx http-server ./dist-storybook",
-    "test": "jest",
-    "test-watch": "jest --watch",
-    "gql-watch-schema": "apollo codegen:generate --config=apollo.config.js --target=typescript --no-addTypename --watch --outputFlat=src/__generatedTypes__",
-    "gql-schema": "apollo codegen:generate --config=apollo.config.js --target=typescript --no-addTypename --outputFlat=src/__generatedTypes__",
-    "make": "yarn plop --plopfile generators/plopfile.js"
+    "prepare": "npx husky install .husky",
+    "prettier-format": "prettier --config .prettierrc.json 'src/**/*.tsx' --write",
+    "serve": "vite preview"
   },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@bitbucket.org/basewd/basevite-blank.git"
+    "url": "git+ssh://git@github.com:PettiSan/fakeTinder.git"
   },
-  "author": "",
   "license": "ISC",
-  "homepage": "https://bitbucket.org/basewd/basevite-blank#readme",
+  "homepage": "https://github.com/PettiSan/fakeTinder",
   "devDependencies": {
-    "@apollo/client": "^3.3.19",
-    "@babel/core": "^7.14.3",
-    "@babel/preset-typescript": "^7.13.0",
-    "@commitlint/cli": "^12.1.1",
-    "@commitlint/config-conventional": "^12.1.1",
-    "@storybook/addon-actions": "^6.2.9",
-    "@storybook/addon-essentials": "^6.2.9",
-    "@storybook/addon-links": "^6.2.9",
-    "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/react": "^6.2.9",
-    "@tailwindcss/aspect-ratio": "^0.2.1",
-    "@tailwindcss/typography": "^0.4.1",
-    "@types/apollo-upload-client": "^14.1.0",
-    "@types/enzyme": "^3.10.8",
-    "@types/jest": "^26.0.23",
-    "@types/react": "^17.0.9",
-    "@types/react-dom": "^17.0.6",
-    "@types/react-helmet": "^6.1.1",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-test-renderer": "^17.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.26.0",
-    "@typescript-eslint/parser": "^4.26.0",
+    "@babel/core": "^7.22.1",
+    "@babel/preset-typescript": "^7.21.5",
+    "@commitlint/cli": "^17.6.5",
+    "@commitlint/config-conventional": "^17.6.5",
+    "@react-spring/web": "^9.7.2",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/react": "^18.2.8",
+    "@types/react-dom": "^18.2.4",
+    "@types/react-helmet": "^6.1.6",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
     "@vitejs/plugin-react-refresh": "^1.3.3",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.1",
-    "apollo": "^2.33.6",
-    "apollo-upload-client": "^15.0.0",
-    "autoprefixer": "^10.2.6",
-    "babel-loader": "^8.2.2",
-    "clsx": "^1.1.1",
-    "concurrently": "^6.2.0",
-    "enzyme": "^3.11.0",
-    "eslint": "^7.27.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^8.1.0",
+    "autoprefixer": "^10.4.14",
+    "babel-loader": "^9.1.2",
+    "clsx": "^1.2.1",
+    "concurrently": "^8.1.0",
+    "eslint": "^8.42.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-loader": "^4.0.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.23.2",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "graphql": "^15.5.3",
-    "husky": "^4.2.5",
-    "jest": "^27.0.4",
-    "jest-css-modules": "^2.1.0",
-    "jest-module-name-mapper": "^0.1.5",
-    "plop": "^2.7.4",
-    "postcss-import": "^14.0.2",
-    "query-string": "^7.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "husky": "^8.0.3",
+    "postcss-import": "^15.1.0",
+    "prettier": "^2.8.8",
+    "query-string": "^8.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-test-renderer": "^17.0.2",
-    "rollup-plugin-visualizer": "^5.5.0",
-    "storybook-css-modules-preset": "^1.1.1",
-    "tailwindcss": "^2.1.4",
-    "ts-jest": "^27.0.2",
-    "tsconfig-paths-webpack-plugin": "^3.5.1",
-    "typescript": "^4.3.2",
-    "use-query-params": "^1.2.2",
-    "vite": "^2.3.6",
-    "vite-tsconfig-paths": "^3.3.13"
+    "react-router-dom": "^6.11.2",
+    "rollup-plugin-visualizer": "^5.9.0",
+    "tailwindcss": "^3.3.2",
+    "tsconfig-paths-webpack-plugin": "^4.0.1",
+    "typescript": "^5.1.3",
+    "vite": "^4.3.9",
+    "vite-tsconfig-paths": "^4.2.0"
   },
   "dependencies": {
     "@types/react-icons": "^3.0.0",
-    "react-icons": "^4.3.1",
-    "react-tinder-card": "^1.4.5"
+    "react-icons": "^4.9.0",
+    "react-tinder-card": "^1.6.2"
   },
   "husky": {
     "hooks": {
+      "pre-commit": "yarn run prettier-format && yarn run lint",
       "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
     }
   },
