@@ -1,6 +1,8 @@
 import { App } from 'App'
-import React from 'react'
-import ReactDom from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './styles/index.css'
 
-ReactDom.render(<App />, document.getElementById('root'))
+const container = document.getElementById('root')
+const root = createRoot(container!)
+
+root.render(<App />)
