@@ -1,29 +1,25 @@
-import React from 'react'
-import { RouteComponentProps } from 'react-router'
+import { lazy } from 'react'
+import { useRoutes } from 'react-router-dom'
 import { PATHS } from './Paths'
 
-const HomePage = React.lazy(() => import('@pages/homePage/HomePage'))
-const MatchesPage = React.lazy(() => import('@pages/matchesPage/MatchesPage'))
-const ChatPage = React.lazy(() => import('@pages/chatPage/ChatPage'))
-const ItsAMatchePage = React.lazy(
+const HomePage = lazy(() => import('@pages/homePage/HomePage'))
+const MatchesPage = lazy(() => import('@pages/matchesPage/MatchesPage'))
+const ChatPage = lazy(() => import('@pages/chatPage/ChatPage'))
+const ItsAMatchePage = lazy(
   () => import('@pages/itsAMatchePage/ItsAMatchePage')
 )
-const ChatWhatIfPage = React.lazy(
+const ChatWhatIfPage = lazy(
   () => import('@pages/chatWhatIfPage/ChatWhatIfPage')
 )
 
-export type Route = {
-  component:
-    | React.ComponentType<RouteComponentProps<any>>
-    | React.ComponentType<any>
-  path: string
-  exact?: boolean
-}
+export function Routes() {
+  const element = useRoutes([
+    { path: PATHS.home, element: <HomePage /> },
+    { path: PATHS.matches, element: <MatchesPage /> },
+    { path: PATHS.chat, element: <ChatPage /> },
+    { path: PATHS.itsAMatche, element: <ItsAMatchePage /> },
+    { path: PATHS.chatWhatIf, element: <ChatWhatIfPage /> },
+  ])
 
-export const ROUTES: Route[] = [
-  { component: HomePage, exact: true, path: PATHS.home },
-  { component: MatchesPage, exact: true, path: PATHS.matches },
-  { component: ChatPage, exact: true, path: PATHS.chat },
-  { component: ItsAMatchePage, exact: true, path: PATHS.itsAMatche },
-  { component: ChatWhatIfPage, exact: true, path: PATHS.chatWhatIf },
-]
+  return <>{element}</>
+}
