@@ -1,5 +1,4 @@
-import React from 'react'
-import Helmet from 'react-helmet'
+import { Helmet } from 'react-helmet'
 
 type SeoProps = {
   title?: string
